+++ packages/block-editor/src/components/block-list/block.js
@@ -410,15 +410,15 @@ function BlockListBlock( {
 			'is-typing': isTypingWithinBlock,
 			'is-focused': isFocusMode && ( isSelected || isParentOfSelectedBlock ),
 			'is-focus-mode': isFocusMode,
+			'has-child-selected': isParentOfSelectedBlock,
 		},
 		className
 	);
 
 	// Determine whether the block has props to apply to the wrapper.
-	let blockWrapperProps = wrapperProps;
 	if ( blockType.getEditWrapperProps ) {
-		blockWrapperProps = {
-			...blockWrapperProps,
+		wrapperProps = {
+			...wrapperProps,
 			...blockType.getEditWrapperProps( attributes ),
 		};
 	}
@@ -473,7 +473,7 @@ function BlockListBlock( {
 			aria-label={ blockLabel }
 			childHandledEvents={ [ 'onDragStart', 'onMouseDown' ] }
 			tagName={ animated.div }
-			{ ...blockWrapperProps }
+			{ ...wrapperProps }
 			style={
 				wrapperProps && wrapperProps.style ?
 					{